requirements.txt
selenium==4.5.0
webdriver-manager==3.8.4
packaging==21.3
bs4==0.0.1
lxml==4.9.1
pandas==1.5.1
openpyxl == 3.0.10
gspread==5.6.2
oauth2client==4.1.3
apiclient==1.0.4
bigquery==0.0.32
python-decouple==3.6